Refactor integration integrator modules
